refactor: isolate jwt from security configuration
