+++ src/main/java/com/academy/e_commerce/config/security/jwt/JwtProvider.java
@@ -1,4 +1,4 @@
-package com.academy.e_commerce.config.security;
+package com.academy.e_commerce.config.security.jwt;
 
 import com.academy.e_commerce.model.UserPrincipal;
 import org.springframework.security.core.Authentication;
